Moved plugins into namespace
