+++ nvim/lua/plugins/toggleterm.lua
@@ -1,0 +1,6 @@
+local status_ok, toggleterm = pcall(require, "toggleterm")
+if not status_ok then
+    return
+end
+
+toggleterm.setup({})
